\name{mills-package}
\alias{mills-package}
\alias{mills}
\docType{package}
\title{
\packageTitle{mills}
}
\description{
\packageDescription{mills}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mills}
\packageIndices{mills}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{mills}

Maintainer: \packageMaintainer{mills}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
